pyproject.toml
[project]
name = "litellm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.74.2",
    "python-dotenv>=1.1.1",
]
